+++ src/components/DatePicker/DateInput.vue
@@ -1,12 +1,16 @@
 <template>
-  <label for="" class="rounded-lg border border-neutral-black p-4">
+  <label
+    for=""
+    class="w-full max-w-[197px] rounded-lg border border-neutral-black bg-neutral-white p-4"
+  >
     <span class="font-regular block text-tiny leading-normal tracking-normal text-neutral-80">{{
       label
     }}</span>
     <input
       type="text"
-      class="tracking-header font-regular block text-body leading-normal outline-none"
+      class="tracking-header font-regular inline-block bg-neutral-transparent text-body leading-normal outline-none"
       :value="date"
+      readonly
   /></label>
 </template>
