+++ src/components/DatePicker/DatePicker.vue
@@ -1,141 +1,36 @@
 <template>
-  <div class="space-y-10">
-    <div class="gap flex items-center justify-between">
-      <ul class="space-y-2">
-        <li class="leading-heading tracking-heading text-h5 font-bold">{{ diffDays }} 晚</li>
-        <li
-          class="font-regular flex items-center justify-center gap-x-2 text-body leading-normal tracking-normal"
-        >
-          <span>{{ formatDate(tempDates?.[0]) }}</span
-          ><span>-</span><span>{{ formatDate(tempDates?.[1]) }}</span>
-        </li>
-      </ul>
-      <div class="flex items-center gap-x-2">
-        <date-input label="入住" :date="formatDate(tempDates?.[0])"></date-input>
-        <date-input label="退房" :date="formatDate(tempDates?.[1])"></date-input>
-      </div>
-    </div>
-    <div class="flex flex-col p-8 desktop:flex-row desktop:gap-x-[60px]">
-      <CalendarPanel
-        key="leftPanel"
-        showArrow="left"
-        :current="leftPanelCurrent"
-        :value="tempDates"
-        @click="onClickArrow"
-        @select="onSelect"
-      />
-      <CalendarPanel
-        key="rightPanel"
-        showArrow="right"
-        :current="rightPanelCurrent"
-        :value="tempDates"
-        @click="onClickArrow"
-        @select="onSelect"
-      />
-    </div>
-    <div class="flex w-full items-center justify-end gap-x-4">
-      <base-button buttonType="ghost" label="" @click="onClear">
-        <p class="text-neutral-80">清除日期</p>
-      </base-button>
-      <base-button label="" @click="onClick">確定日期</base-button>
-    </div>
+  <div ref="datePickerRef" class="flex items-center gap-x-2" @click="showModal = true">
+    <date-input label="入住"></date-input>
+    <date-input label="退房"></date-input>
   </div>
+  <Teleport to="body">
+    <date-picker-modal
+      v-model:show="showModal"
+      :top="position.top"
+      :right="position.right"
+    ></date-picker-modal>
+  </Teleport>
 </template>
 
 <script setup lang="ts">
-import { ref, computed, toRefs } from 'vue';
-import CalendarPanel from './CalendarPanel.vue';
+import { ref, watch, reactive } from 'vue';
 import DateInput from './DateInput.vue';
-import BaseButton from '../BaseButton.vue';
-import { diffDate, curryFormatDate } from './utils';
+import DatePickerModal from './DatePickerModal.vue';
+import { useBoolean, useRect } from '@/hooks';
 
-type Props = Partial<{
-  value: [number, number] | null;
-}>;
+const datePickerRef = ref<null | HTMLDivElement>(null);
+const { bool: showModal } = useBoolean(false);
+const position = reactive({ top: 0, right: 0 });
+const datePickerRect = useRect(datePickerRef);
 
-const props = withDefaults(defineProps<Props>(), {
-  value: null,
-});
-const emits = defineEmits<{
-  'update:value': [value: [number, number] | null];
-}>();
-const { value } = toRefs(props);
-
-const tempDates = ref<[] | [number] | [number, number] | null>(value.value);
-
-const leftPanelCurrent = ref(new Date());
-const rightPanelCurrent = computed(() => {
-  const date = new Date(leftPanelCurrent.value);
-  return new Date(date.setMonth(date.getMonth() + 1));
-});
-
-const diffDays = computed<number>(() => {
-  if (tempDates.value && tempDates.value.length === 2) {
-    const [start, end] = tempDates.value;
-    return diffDate(end, start, 'day');
-  }
-  return 0;
-});
-
-function onClickArrow(direction: 'left' | 'right') {
-  if (direction === 'left') {
-    onPreMonth();
-  }
-  if (direction === 'right') {
-    onNextMonth();
-  }
-}
-
-function onSelect(date: number) {
-  if (tempDates.value?.length === 2) {
-    tempDates.value = [];
-  }
-  if (tempDates.value === null) {
-    tempDates.value = [];
-  }
-  switch (tempDates.value?.length) {
-    case 0: {
-      tempDates.value = [date];
-      break;
-    }
-    case 1: {
-      const [d] = tempDates.value;
-      if (d > date) {
-        tempDates.value = [date, d];
-      } else {
-        tempDates.value = [d, date];
-      }
-      break;
+watch(showModal, (show) => {
+  if (show) {
+    if (datePickerRect.value) {
+      position.top = datePickerRect.value.top - 32;
+      position.right = innerWidth - datePickerRect.value.right + 32 + 17;
     }
   }
-}
-
-function onPreMonth() {
-  leftPanelCurrent.value = new Date(
-    leftPanelCurrent.value.setMonth(leftPanelCurrent.value.getMonth() - 1),
-  );
-}
-
-function onNextMonth() {
-  leftPanelCurrent.value = new Date(
-    leftPanelCurrent.value.setMonth(leftPanelCurrent.value.getMonth() + 1),
-  );
-}
-
-function onClick() {
-  if (tempDates.value && tempDates.value.length === 2) {
-    emits('update:value', [tempDates.value[0], tempDates.value[1]]);
-  }
-}
-
-function onClear() {
-  tempDates.value = null;
-  emits('update:value', tempDates.value);
-}
-
-function formatDate(date?: number): string {
-  return (date && curryFormatDate('YYYY/ MM / DD')(date)) || '-';
-}
+});
 </script>
 
 <style scoped></style>
